data-standardization/data-standard_and_normal/requirements.txt
# Core Data Processing
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0
pytz==2024.1
# Config + Validation
pydantic==2.7.1
pydantic-settings==2.2.1
PyYAML==6.0.1
python-dotenv==1.0.1

# API Layer (serverless / Vercel compatible)
fastapi==0.111.0
uvicorn==0.30.0
rapidfuzz==3.9.0
unidecode==1.3.8
forex-python==1.8
loguru==0.7.2
rich==13.7.1
pytest==8.2.0
pytest-cov==5.0.0
black==24.4.2
isort==5.13.2
mypy==1.10.0
pyarrow==16.1.0
fastapi
uvicorn
pandas
pytest
python-dotenv
fastapi
pandas
mangum
python-multipart